requirements.txt
yt-dlp
pydub
python-dotenv
groq
requests
openai-whisper
torch
